testing_helpers clean up
